Improvments in the frontend: reloading / deleting / vnc.

+++ Makefile
@@ -1,14 +1,15 @@
 OBJ=$(shell ls src/*.erl | sed -e 's/\.erl$$/.beam/' | sed -e 's/^src/ebin/g') $(shell ls src/*.app.src | sed -e 's/\.src$$//g' | sed -e 's/^src/ebin/g')
 DEPS=$(shell cat rebar.config  |sed -e 's/%.*//'| sed -e '/{\(\w\+\), [^,]\+, {\w\+, [^,]\+, {[^,]\+, [^}]\+}}},\?/!d' | sed -e 's;{\(\w\+\), [^,]\+, {\w\+, [^,]\+, {[^,]\+, [^}]\+}}},\?;deps/\1/rebar.config;')
 ERL=erl
-PA=ebin deps/*/ebin
+PA=$(shell pwd)/ebin
+ERL_LIBS=`pwd`/deps/
 REBAR=./rebar
 APP_NAME=wiggle
 
 all: $(OBJ) $(DEPS)
 
 rel: all FORCE
-	-rm -r rel/wiggle
+	-rm -r rel/$(APP_NAME)
 	cd rel; ../rebar generate
 echo:
 	echo $(DEPS)
@@ -17,8 +18,8 @@ tar: rel
 	cd rel; tar jcvf $(APP_NAME).tar.bz2 $(APP_NAME)
 
 clean: FORCE
-	-rm -r *.beam ebin
-	-rm erl_crash.dump
+	$(REBAR) clean
+	-rm *.beam erl_crash.dump
 	-rm -r rel/$(APP_NAME)
 	-rm rel/$(APP_NAME).tar.bz2
 
@@ -32,7 +33,7 @@ ebin/%.beam: src/%.erl
 	$(REBAR) compile
 
 shell: all
-	$(ERL) -pa $(PA) -configig standalone.configig
-	-rm *.beam
+	ERL_LIBS="$(ERL_LIBS)" $(ERL) -pa $(PA) -config standalone
+	[ -f *.beam ] && rm *.beam || true
 
 FORCE:
